AEA-3899 Leftover function moved into fhirUtils. Made enrichPrescriptions closer to common fhir code.
